requirements.txt
# Core Flask
Flask==3.1.2
Werkzeug==3.1.3
wtforms==3.1.2

# Database
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
psycopg2-binary==2.9.10
SQLAlchemy==1.4.54

# Authentication & Security
Flask-JWT-Extended==4.5.3
PyJWT==2.8.0
cryptography==41.0.4
pyotp==2.8.0
Flask-Talisman==1.1.0

# API & Utilities
Flask-CORS==4.0.0
Flask-Limiter==3.5.0
Flask-Compress==1.14
python-dotenv==1.0.0
python-slugify==8.0.1



# Task Queue
celery==5.3.4

# Optional (remove if not needed)
Flask-Mail==0.9.1
Flask-Login==0.6.3


# Redis & Caching
Flask-Caching==2.1.0
redis==5.0.1

requests==2.32.5







